requirements.txt
fastapi
uvicorn[standard]
supabase
python-jose[cryptography]
httpx
python-dotenv
